MODORDSTOR-222. Make /orders-storage/po-line-number endpoint return list of poLine numbers
