+++ ramls/po-line-number.raml
@@ -4,11 +4,11 @@ baseUri: http://github.com/folio-org/mod-orders-storage
 version: v1
 
 documentation:
-  - title: Purchase Order Line Number
-    content: <b>API used to manage Purchase Order Line number.  This API is intended for internal use only</b>
+  - title: Purchase Order Line Numbers
+    content: <b>API used to manage Purchase Order Line numbers.  This API is intended for internal use only</b>
 
 types:
-  po-line-number: !include acq-models/common/schemas/sequence_number.json
+  po-line-number: !include acq-models/common/schemas/sequence_numbers.json
 
 traits:
     language: !include raml-util/traits/language.raml
@@ -16,13 +16,18 @@ traits:
 /orders-storage/po-line-number:
   get:
     is: [language]
-    description: Get purchase order line number
+    description: Get purchase order line numbers
     queryParameters:
       purchaseOrderId:
         description: Purchase Order Id
         type:        string
         required:    true
         example:     "8ad4b87b-9b47-4199-b0c3-5480745c6b41"
+      poLineNumbers:
+        description: Quantity of the PO line numbers
+        type:        integer
+        default:     1
+        example:     1
     responses:
       200:
         body:
